Created new host image review screen to search for question images
This adds an 'image' and 'image_url' attribute to the question
class. When the board is generated, any question with a link in
it on j-archive, will be marked 'True' for image, as it may
require an image.

When a question is selected on HostQuestionCard.mousePressEvent()
if question.image is True, it will load image review screen instead
of question.

The image review screen will show the host the question text
and an image search box. The host can enter in queries to wikimedia
or urls. At start, the result of the wikimedia search for question
answer is shown.

Two new keyboard event listeners are activated on this screen.
Shift accepts the current image, control sets the active question's
image value to False. The screen is then re-routed to the classic
question loading route.

Question loading now checks if a question has an image_url and 'image'
is set to True. If so, it will attempt to load the image.

TODO:
- make it so images can be loaded from url or locally
- change the name HostImageQuestionWidget
- make HostImageQuestionWidget look better
- autoscale images better
- set up setting 'image_url' at board resolution if the images
    exist locally
- get rid of useless accept and reject buttons
- make keys to approve/regect images intuitive